fix: defer projectless chat creation
