+++ packages/client/app/pages/chats/index.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+import ChatSessionShell from '../../components/ChatSessionShell.vue'
+</script>
+
+<template>
+  <ChatSessionShell />
+</template>
